+++ app/javascript/vue/components/ui/Date/DateToday.vue
@@ -2,35 +2,27 @@
   <v-btn
     color="primary"
     medium
-    @click="setTime"
+    @click="emitActualDate"
   >
-    Now
+    Today
   </v-btn>
 </template>
 
 <script setup>
-import { convertToTwoDigits } from 'helpers/numbers'
+
 import VBtn from 'components/ui/VBtn/index.vue'
 
 const emit = defineEmits([
   'update:day',
   'update:month',
-  'update:year',
-  'update:time'
+  'update:year'
 ])
 
-const setTime = () => {
+const emitActualDate = () => {
   const today = new Date()
-  const time = [
-    convertToTwoDigits(today.getHours()),
-    convertToTwoDigits(today.getMinutes()),
-    convertToTwoDigits(today.getSeconds())
-  ].join(':')
 
-  emit('update:time', time)
   emit('update:day', today.getDate())
   emit('update:month', today.getMonth() + 1)
   emit('update:year', today.getFullYear())
 }
-
 </script>