+++ app/javascript/vue/helpers/numbers.js
@@ -1,0 +1,6 @@
+export function convertToTwoDigits (number) {
+  return Number(number).toLocaleString('en-US', {
+    minimumIntegerDigits: 2,
+    useGrouping: false
+  })
+}