+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 - Protocols for Observations [#2889]
 - Increment tripcode in New collecting event task [#2441]
+- Now and today buttons for time/date made [#2888]
 
 ### Changed
 - Refactor Uniquify People task. Added improvements [#2858]
@@ -21,6 +22,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#2889]: https://github.com/SpeciesFileGroup/taxonworks/issues/2889
 [#2441]: https://github.com/SpeciesFileGroup/taxonworks/issues/2441
 [#2858]: https://github.com/SpeciesFileGroup/taxonworks/issues/2858
+[#2888]: https://github.com/SpeciesFileGroup/taxonworks/issues/2888
 [#2892]: https://github.com/SpeciesFileGroup/taxonworks/issues/2892
 
 ## [0.24.1] - 2022-04-04